Quiz Button Color Change
Modified answer script to change the color of quiz game buttons when the player answers correctly. On a correct answer, the answer script will start a coroutine which changes the color of the button to green, waits three seconds, changes the button color back to white, and calls the correct answer function of the quiz manager.